requirements.txt
# 自动钓鱼项目依赖列表 v2.3.0

# 核心框架
opencv-python>=4.8.0        # OpenCV图像处理库
numpy>=1.24.0               # 数值计算库
pillow>=9.0.0               # 图像处理库

# 深度学习框架
ultralytics>=8.0.0          # YOLOv8/YOLO11模型库
torch>=2.0.0                # PyTorch深度学习框架
torchvision>=0.15.0          # PyTorch视觉库

# 屏幕操作和控制
pyautogui>=0.9.54           # 屏幕截图和鼠标键盘控制
keyboard>=0.13.5            # 键盘监听
mouse>=0.7.1                # 鼠标操作
mss>=9.0.1                  # 快速屏幕截图（性能优化）

# 文字识别（OCR）
pytesseract>=0.3.10         # Tesseract OCR Python接口

# 数据处理和配置
pyyaml>=6.0                 # YAML文件处理
pandas>=2.0.0               # 数据分析和处理
pathlib2>=2.3.7            # 路径处理（Python < 3.4兼容）

# GUI界面
tkinter                     # GUI库（Python内置，通常不需要安装）
customtkinter>=5.2.0        # 现代化的tkinter主题扩展

# 日志和系统监控
psutil>=5.9.0               # 系统资源监控
tqdm>=4.65.0                # 进度条显示

# 开发和调试工具
matplotlib>=3.7.0           # 图表绘制和可视化
seaborn>=0.12.0             # 统计图表

# 性能优化（可选）
numba>=0.57.0               # JIT编译加速

# 网络请求（如果需要在线功能）
requests>=2.31.0            # HTTP请求库

# 机器学习辅助工具
scikit-learn>=1.3.0         # 机器学习库

# ===== 安装说明 =====
# 1. 基本安装：pip install -r requirements.txt
# 2. GPU支持：需要安装对应CUDA版本的PyTorch
#    - CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#    - CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# 3. 在某些Linux系统上可能需要额外安装：
#    - sudo apt-get install python3-tk python3-dev
# 4. 如果使用conda环境：
#    - conda install tk opencv numpy matplotlib
# 5. Tesseract OCR需要单独安装系统级组件：
#    - Windows: 已配置路径 D:\Python\tool\Tesseract-OCR
#    - Linux: sudo apt-get install tesseract-ocr tesseract-ocr-chi-sim
#    - macOS: brew install tesseract tesseract-lang

# ===== 系统要求 =====
# - Python 3.8+
# - Windows 10/11 (推荐) / Linux / macOS
# - 内存: 8GB+ (推荐16GB+)
# - 显卡: NVIDIA GPU with CUDA支持 (可选，用于加速)
# - 硬盘: 至少5GB可用空间

# ===== 开发依赖 =====
# 以下包仅在开发环境中需要，生产环境可不安装
# pytest>=7.0.0             # 单元测试框架
# black>=22.0.0              # 代码格式化工具  
# flake8>=5.0.0              # 代码质量检查
# mypy>=1.0.0                # 类型检查工具 